Refactor limits-checking
Drop the use of the term "requirements" in favour of just "limits".
Support for limits is now in modules defined under lib/support,
with a new LimitsContext class to handle the needed dependencies.
The module lib/node/limits.js contains now only the functions that apply
different limit policies in use.