Deprecated SQL Function analysis
It allows calling a PosgreSQL function to create and populate an
analysis cache table. The function receives the name of the table
to create or populate, and the operation it has to perform, either
`create` or `populate`.

The table has to have at least:
 - A `the_geom Geometry(4326)` column.
 - A `cartodb_id numeric` column.

An example of this kind of functions can be found in the example
and in the tests included in this commit.