Bind node itself in beforeCreate function for better semantics
It's easier to understand what you are affecting when you work
with `this` instead with a pass by reference `node` variable.

This will align with other hooks like one for being able to execute
an arbitrary query before creating the node to ensure/validate some
metadata is present in the created node.