Using a partial index over the geometry column of the cached table
In CARTO, user datasets have their geometries ready for rendering:
they are already projected and have an appropriate index.
On the opposite side, Camshaft's cached tables are not ready for
rendering: missing the adequate projection and the index.

Some numbers on this solution:
https://gist.github.com/rochoa/cf0f512cdad666f62ee5b8930cf4db94.

This is an alternative to and closes #200 and in consequence,
it also closes #201.